Module Opentelemetry_client

module Bounded_queue : sig ... end

Interface for a thread-safe, bounded queue.

module Common_ : sig ... end

module Consumer : sig ... end

Consumer that accepts items from a bounded queue and processes them.

module Emitter_combine : sig ... end

Combine multiple emitters into one.

module Emitter_limit_interval : sig ... end

Limit frequency at which the emitter emits.

module Emitter_sample : sig ... end

module Export_error : sig ... end

Error that can occur during export

module Exporter_combine : sig ... end

Combine multiple exporters into one

module Exporter_config : sig ... end

Constructing and managing the configuration common to many (most?) HTTP-based clients.

module Exporter_debug : sig ... end

Basic debug exporter, prints signals on stdout/stderr/...

module Exporter_queued : sig ... end

Build an exporter from a queue and a consumer.

module Exporter_stdout : sig ... end

A simple exporter that prints on stdout.

module Generic_consumer : sig ... end

A consumer: pulls signals from a queue, sends them somewhere else

module Generic_consumer_exporter : sig ... end

A consumer that just calls another exporter.

module Generic_http_consumer : sig ... end

module Generic_io : sig ... end

Generic IO monad.

module Generic_notifier : sig ... end

Generic notifier (used to signal when a bounded queue is empty)

module Http_config : sig ... end

module Resource_signal : sig ... end

Constructing and managing OTel signals at the resource (batch) level

module Sampler : sig ... end

Basic random sampling.

module Self_trace : sig ... end

Mini tracing module for OTEL itself.

module Util_net_backoff : sig ... end

Backoff behavior in case of errors

module Util_resources : sig ... end

Group signals into resource_xxx objects