Module Opentelemetry_client.Consumer

Consumer that accepts items from a bounded queue and processes them.

type t = {

active : unit -> Opentelemetry_util.Aswitch.t;

shutdown : unit -> unit;
(*
Shutdown the consumer as soon as possible. active will be turned off once the consumer is fully shut down.
*)

tick : unit -> unit;
(*
Regularly called, eg to emit metrics, check timeouts, etc. Must be thread safe.
*)

self_metrics : clock:Opentelemetry_client.Common_.OTEL.Clock.t -> unit -> Opentelemetry_client.Common_.OTEL.Metrics.t list;
(*
Self observing metrics
*)

}

A consumer for signals of type 'a

type consumer = t

val active : t -> Opentelemetry_util.Aswitch.t

val shutdown : t -> unit

val self_metrics : clock:Opentelemetry_client.Common_.OTEL.Clock.t -> t -> Opentelemetry_client.Common_.OTEL.Metrics.t list

val on_stop : t -> (unit -> unit) -> unit

on_stop e f calls f() when e stops, or now if it's already stopped

module Builder : sig ... end

type any_signal_l_builder = Opentelemetry_client.Common_.OTEL.Any_signal_l.t Builder.t

type resource_signal_builder = Resource_signal.t Builder.t

The type that's useful for HTTP backends