Module Consumer.Builder

type 'a t = {

start_consuming : 'a Bounded_queue.Recv.t -> consumer;

}

A builder that will create a consumer for a given queue, start the consumer so it starts consuming from the queue, and return the consumer.

val start_consuming : 'a t -> 'a Bounded_queue.Recv.t -> consumer

val map : ('a -> 'b) -> 'b t -> 'a t