Module Opentelemetry_client.Emitter_combine

Combine multiple emitters into one.

type closing_behavior = [

| `Close_when_all_closed

| `Close_when_one_closed

]

When to close the combined emitter:

`Close_when_all_closed: closed when all the emitters that are combined are closed

`Close_when_one_closed: closed as soon as one of the emitters is closed

val combine_l : ?closing:closing_behavior -> 'a Opentelemetry_emitter.Emitter.t list -> 'a Opentelemetry_emitter.Emitter.t

val combine : 'a Opentelemetry_emitter.Emitter.t -> 'a Opentelemetry_emitter.Emitter.t -> 'a Opentelemetry_emitter.Emitter.t