Module Generic_consumer.Make

Parameters

module IO : IO

module Notifier : Generic_notifier.S with type 'a IO.t = 'a IO.t

module Sender : SENDER with type 'a IO.t = 'a IO.t

Signature

val consumer : sender_config:Sender.config -> n_workers:int -> ticker_task:float option -> ?on_tick:(unit -> unit) -> unit -> Consumer.any_signal_l_builder

Make a consumer builder, ie. a builder function that will take a bounded queue of signals, and start a consumer to process these signals and send them somewhere using HTTP.