Module Opentelemetry_client.Generic_consumer_exporter

A consumer that just calls another exporter.

This is useful to introduce queueing behavior using Exporter_queued, but simply forwarding to another (presumably non-queue) exporter.

It is generic because we need some sort of threading/concurrency to run the consumer.

module type IO = Generic_io.S_WITH_CONCURRENCY

module Make (IO : IO) (Notifier : Generic_notifier.S with type 'a IO.t = 'a IO.t) : sig ... end