Module Generic_http_consumer.Make

Parameters

module IO : IO

module Notifier : Generic_notifier.S with type 'a IO.t = 'a IO.t

module Httpc : HTTPC with type 'a IO.t = 'a IO.t

Signature

val consumer : ?override_n_workers:int -> ticker_task:float option -> ?on_tick:(unit -> unit) -> config:Http_config.t -> unit -> Consumer.any_signal_l_builder

Make a consumer builder, ie. a builder function that will take a bounded queue of signals, and start a consumer to process these signals and send them somewhere using HTTP.

parameter ticker_task
controls whether we start a task to call tick at the given interval in seconds, or None to not start such a task at all.