Module Opentelemetry_client.Resource_signal

Constructing and managing OTel signals at the resource (batch) level

type t =

| Traces of Opentelemetry_proto.Trace.resource_spans list

| Metrics of Opentelemetry_proto.Metrics.resource_metrics list

| Logs of Opentelemetry_proto.Logs.resource_logs list

The type of signals

This is not the principle type of signals from the perspective of what gets encoded and sent via protocl buffers, but it is the principle type that collector clients needs to reason about.

val pp : Stdlib.Format.formatter -> t -> unit

val of_logs : ?service_name:string -> ?attrs:Opentelemetry_client.Common_.OTEL.Key_value.t list -> Opentelemetry_client.Common_.Proto.Logs.log_record list -> t

val of_logs_or_empty : ?service_name:string -> ?attrs:Opentelemetry_client.Common_.OTEL.Key_value.t list -> Opentelemetry_client.Common_.Proto.Logs.log_record list -> t list

val of_spans : ?service_name:string -> ?attrs:Opentelemetry_client.Common_.OTEL.Key_value.t list -> Opentelemetry_client.Common_.OTEL.Span.t list -> t

val of_spans_or_empty : ?service_name:string -> ?attrs:Opentelemetry_client.Common_.OTEL.Key_value.t list -> Opentelemetry_client.Common_.OTEL.Span.t list -> t list

val of_metrics : ?service_name:string -> ?attrs:Opentelemetry_client.Common_.OTEL.Key_value.t list -> Opentelemetry_client.Common_.Proto.Metrics.metric list -> t

val of_metrics_or_empty : ?service_name:string -> ?attrs:Opentelemetry_client.Common_.OTEL.Key_value.t list -> Opentelemetry_client.Common_.Proto.Metrics.metric list -> t list

val of_signal_l : ?service_name:string -> ?attrs:Opentelemetry_client.Common_.OTEL.Key_value.t list -> Opentelemetry_client.Common_.OTEL.Any_signal_l.t -> t

val to_traces : t -> Opentelemetry_proto.Trace.resource_spans list option

val to_metrics : t -> Opentelemetry_proto.Metrics.resource_metrics list option

val to_logs : t -> Opentelemetry_proto.Logs.resource_logs list option

val is_traces : t -> bool

val is_metrics : t -> bool

val is_logs : t -> bool

type protocol = Exporter_config.protocol =

| Http_protobuf

| Http_json

module Encode : sig ... end

Encode signals to protobuf or JSON encoded strings, ready to be sent over the wire

module Decode : sig ... end

Decode signals from protobuf encoded strings, received over the wire

module Pp : sig ... end